refactor: `mutex_utility` を統合して廃止
